// src/shared/queries/types/insertThemes.ts
/* tslint:disable */
/* eslint-disable */
// This file was automatically generated and should not be edited.

import { InputThemeSettings } from "./../../../../types/globalTypes";

// ====================================================
// GraphQL mutation operation: insertThemes
// ====================================================

export interface insertThemes {
  insertThemes: boolean;
}

export interface insertThemesVariables {
  name: string;
  settings: InputThemeSettings[];
}